Rename screens back to actions.
