+++ bin/app
@@ -10,10 +10,10 @@ require 'db/reader'
 require 'db/writer'
 require 'db/db'
 
-require 'screens/lister'
-require 'screens/creator'
-require 'screens/finder'
-require 'screens/updater'
+require 'actions/lister'
+require 'actions/creator'
+require 'actions/finder'
+require 'actions/updater'
 
 require 'menu'
 require 'conman'
