+++ lib/actions/lister.rb
